% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{census_inhabitants}
\alias{census_inhabitants}
\title{Census 2011 Inhabitants Data}
\format{
A data frame/simple feature collection with 361,478 rows and 13 columns:
\describe{
  \item{inspid1km}{INSPIRE 1 sqkm identifier}
  \item{inhabitants}{Inhabitants count}
  \item{geometry}{Point geometry column (\code{EPSG:3035})}
  \item{ags_YEAR}{German municipality identifier for the corresponding year ("Allgemeiner Gemeindeschlüssel")}
}
}
\source{
https://www.zensus2011.de
}
\usage{
census_inhabitants
}
\description{
Census 2011 Inhabitants Data
}
\keyword{datasets}
