% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R
\name{draw_sample}
\alias{draw_sample}
\title{Draw a Sample Based on a Defined Sample Frame}
\usage{
draw_sample(sample_frame = NULL, randomize = TRUE, verbose = TRUE)
}
\arguments{
\item{sample_frame}{A custom sample frame, if provided, overrides the default.}

\item{randomize}{Logical. If TRUE (default), shuffles the final sample.}

\item{verbose}{Show output from function (default TRUE)}
}
\value{
A tibble with selected INSPIRE grid cell identifiers.
}
\description{
This function selects a sample of INSPIRE grid cells from a pre-defined sample frame,
considering population data for weighting. It supports randomization.
}
\examples{
library(geosynth)

data("fake_survey_coordinates")

sample_frame <-
 geosynth::create_sample_frame(
   .data = fake_survey_coordinates,
   year = "2024",
   geo_unit = "regiostar17",
   inhabitants_threshold = 10000,
   minimum_sample_points = 10
 )

synthetic_sample <- geosynth::draw_sample(sample_frame = sample_frame)

}
