% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R
\name{draw_sample}
\alias{draw_sample}
\title{Draw a Sample Based on a Defined Sample Frame}
\usage{
draw_sample(sample_frame = NULL, year = NULL, randomize = TRUE)
}
\arguments{
\item{sample_frame}{A custom sample frame, if provided, overrides the default.}

\item{randomize}{Logical. If TRUE (default), shuffles the final sample.}

\item{custom_year}{A custom year for retrieving municipality and census data. If NULL,
the function extracts the year from `survey`.}
}
\value{
A tibble with selected INSPIRE grid cell identifiers.
}
\description{
This function selects a sample of INSPIRE grid cells from a pre-defined sample frame,
considering population data for weighting. It supports randomization.
}
