% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fake_survey_coordinates}
\alias{fake_survey_coordinates}
\title{Fake Survey Coordinates}
\format{
A data frame/simple feature collection with 3,004 rows and 3 columns:
\describe{
  \item{id}{Fake respondent identifier}
  \item{ags}{8-digit German municipality identifier ("Allgemeiner Gemeindeschlüssel")}
  \item{geometry}{Point geometry column (\code{EPSG:3035})}
}
}
\source{
Created using \code{geosynth:::create_fake_survey_coordinates()}
}
\usage{
fake_survey_coordinates
}
\description{
Fake survey coordinates that are not even synthetic and only serve
demonstration purposes
}
\keyword{datasets}
