% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mun_2012}
\alias{mun_2012}
\alias{mun_2013}
\alias{mun_2014}
\alias{mun_2015}
\alias{mun_2016}
\alias{mun_2017}
\alias{mun_2018}
\alias{mun_2019}
\alias{mun_2020}
\alias{mun_2021}
\alias{mun_2022}
\alias{mun_2023}
\alias{mun_2024}
\title{Municipality Level Data}
\format{
A data frame with ~11,000 rows (depending on the year) and 6 columns:
\describe{
  \item{lan}{2-digit German federal state identifier}
  \item{ags_YEAR}{8-digit German municipality identifier ("Allgemeiner Gemeindeschlüssel")}
  \item{gkpol}{1-digit political municipality size class identifier ("Politische Gemeindegrößenklasse"), see \url{https://www.bbsr.bund.de/BBSR/DE/forschung/raumbeobachtung/Raumabgrenzungen/deutschland/gemeinden/StadtGemeindetyp/StadtGemeindetyp.html}}
  \item{regiostar7}{2-digit summarized regional spatial typology statistic identifier ("Regionalstatistische Raumtypologie, Zusammengefasster regionalstatischer Raumtyp"), see \url{https://www.bmv.de/SharedDocs/DE/Artikel/G/regionalstatistische-raumtypologie.html}}
  \item{regiostar17}{3-digit regional spatial statistic type identifier ("Regionalstatistische Raumtypologie, Regionalstatischer Raumtyp"), see \url{https://www.bmv.de/SharedDocs/DE/Artikel/G/regionalstatistische-raumtypologie.html}}
  \item{inhabitants}{Inhabitants count in absolute numbers}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11093 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11105 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11116 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11121 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11088 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11083 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11043 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11005 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10999 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10992 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10988 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10976 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10954 rows and 6 columns.
}
\source{
https://opendata-esridech.hub.arcgis.com

https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales/Gemeindeverzeichnis/Administrativ/15-regiostar.html

https://www.bbsr.bund.de/BBSR/DE/forschung/raumbeobachtung/Raumabgrenzungen/deutschland/gemeinden/StadtGemeindetyp/StadtGemeindetyp.html
}
\usage{
mun_2012

mun_2013

mun_2014

mun_2015

mun_2016

mun_2017

mun_2018

mun_2019

mun_2020

mun_2021

mun_2022

mun_2023

mun_2024
}
\description{
These datasets contain municipality-level administrative and regional
classification variables for Germany for multiple years. They are used
as lookup tables for spatial enrichment and sample frame construction
within the `geosynth` package.
}
\keyword{datasets}
