% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_min_distance.R
\name{shuffle_min_distance}
\alias{shuffle_min_distance}
\title{Shuffle Synthetic Data to Maintain Minimum Distance}
\usage{
shuffle_min_distance(
  synthetic_data,
  original_data,
  min_km = 50,
  max_tries = 1000,
  verbose = TRUE
)
}
\arguments{
\item{synthetic_data}{An `sf` object with the synthetic spatial data to shuffle.}

\item{original_data}{An `sf` object containing the original spatial data.}

\item{min_km}{Minimum allowed distance (in kilometers) between original
and corresponding synthetic points. Default is 50.}

\item{max_tries}{Maximum number of attempts to achieve the minimum distance
condition. Default is 10,000.}

\item{verbose}{Show output from function (default TRUE)}
}
\value{
An `sf` object where the synthetic data has been shuffled such that
  all points are at least `min_km` apart from the original data.
}
\description{
This function attempts to shuffle rows in a synthetic dataset so that
each point is at least `min_km` kilometers away from its corresponding
point in the original dataset. It performs iterative reassignments
for problematic points until either the distance condition is satisfied
or `max_tries` is reached.
}
\examples{
library(geosynth)

data("fake_survey_coordinates")

sample_frame <-
 geosynth::create_sample_frame(
   .data = fake_survey_coordinates,
   year = "2024",
   geo_unit = "regiostar17",
   inhabitants_threshold = 10000,
   minimum_sample_points = 10
 )

synthetic_sample <- geosynth::draw_sample(sample_frame = sample_frame)

synthetic_shuffled <-
  geosynth::shuffle_min_distance(
    synthetic_data = synthetic_sample,
    original_data = fake_survey_coordinates,
    min_km = 5   # kilometers
  )
}
